requirements.txt
streamlit 
moviepy 
openai-whisper 
langchain-core 
langchain-community 
faiss-cpu 
sentence-transformers 
python-dotenv 
langchain-groq